Create/update getUser, patchEmail, deleteUser

+++ src/controllers/user-controller.js
@@ -1,9 +1,11 @@
+import bcrypt from 'bcrypt';
 import User from '../models/User.js';
 
 export const getUser = async (req, res) => {
     const userId = req.params.id;
     if (!userId) {
         res.json({ error: 'No user ID provided.' });
+        return;
     }
     try {
         const {
@@ -29,57 +31,60 @@ export const getUser = async (req, res) => {
     }
 };
 
-/*
-    Make sure the body of the request is of type JSON with values:
-    BE CAREFUL OF THE CASE SENSITIVITY
-    {
-        {
-            "firstName": "firstName",
-            "lastName": "lastName",
-            "email": "email@email.com",
-            "password": "password"
-        }
+export const patchEmail = async (req, res) => {
+    const { userId, newEmail, password } = req.body;
+    const emailRegExp = new RegExp(
+        /[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}/g
+    );
+    if (!userId) {
+        res.status(400).json({ error: 'No user ID provided.' });
+        return;
     }
-*/
-
-export const postUser = async (req, res) => {
-    try {
-        const user = await User.create({
-            firstName: req.body.firstName,
-            lastName: req.body.lastName,
-            email: req.body.email,
-            password: req.body.password,
-            usertype: req.body.usertype,
+    if (!newEmail || !emailRegExp.test(newEmail)) {
+        res.status(400).json({
+            error: 'Please enter a valid email address.',
         });
-        res.json({ user });
-    } catch (err) {
-        res.json({ error: err });
+        return;
     }
-};
-
-export const patchUser = async (req, res) => {
+    if (!password) {
+        res.status(400).json({ error: 'No password provided.' });
+        return;
+    }
     try {
-        const user = await User.update(
-            {
-                firstName: req.body.firstName,
-                lastName: req.body.lastName,
-                email: req.body.email,
-                password: req.body.password,
-                usertype: req.body.usertype,
+        const user = await User.findOne({
+            where: {
+                id: userId,
             },
-            { where: { id: req.params.id } }
-        );
-        res.json({ user });
+        });
+        if (!user) {
+            res.status(400).json({ error: 'Incorrect user ID.' });
+            return;
+        }
+        const passwordCorrect = await bcrypt.compare(password, user.password);
+        if (!passwordCorrect) {
+            res.status(401).json({ error: 'Incorrect password.' });
+            return;
+        }
+        user.email = newEmail;
+        await user.save();
+        res.sendStatus(200);
     } catch (err) {
-        res.json({ error: err });
+        res.status(400).json({ error: '' });
     }
 };
 
 export const deleteUser = async (req, res) => {
+    if (!req.params.id) {
+        res.status(400).json({ error: 'No user ID provided.' });
+        return;
+    }
+    if (req.params.id !== req.userId) {
+        res.status(401).json({ error: 'Unauthorised.' });
+    }
     try {
         await User.destroy({ where: { id: req.params.id } });
         res.sendStatus(200);
     } catch (err) {
-        res.json({ error: err });
+        res.status(400).json({ error: 'Incorrect user ID.' });
     }
 };
